fix: 화면 회전 시 data가 초기화 되는 문제 해결 - 데이터만 관리하는 파일을 따로 생성하여 기존 코드와 분리 - ViewModel을 생성하여 recomposition이 발생해도 데이터 값이 유지하도록 변경 -NavigationItem => MainScreen으로 파일 이름 변경
