+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,8 @@
-#Tue Mar 11 19:20:44 KST 2025
+#Sat Mar 22 19:30:35 KST 2025
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 distributionUrl=https\://services.gradle.org/distributions/gradle-8.10.2-bin.zip
+networkTimeout=10000
+validateDistributionUrl=true
 zipStoreBase=GRADLE_USER_HOME
 zipStorePath=wrapper/dists
